backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
google-generativeai==0.8.3
